pyproject.toml
[tool.poetry]
name = "trip-planner-crew"
version = "0.1.0"
description = "Simple trip planner"
authors = ["Shawn Vo"]

[tool.poetry.dependencies]
python =  ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^0.28.8"}
python-dotenv = "1.0.0"
langchain-community = "^0.0.34"
amadeus = "^10.0.0"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"